worked on finishing darkmode Feature. extracted function to external mod
