+++ task-app/src/App.scss
@@ -32,7 +32,7 @@ $deskLight:url('./images/bg-desktop-light.jpg');
 @mixin darkMode($bgColor:$nightVDesBlue){
   background-color:$bgColor;
   color:$dayVLightGray;
-  transition: 2s;
+
 }
 
 
@@ -66,8 +66,10 @@ body{
   padding:0;
   box-sizing: border-box;
   background-color:$dayVLightGray;
-  transition: 2s;
+
   font-family: $fontFamily;
+
+
 }
 
 .App{
@@ -83,6 +85,7 @@ body{
     color: $dayVLightGray;
     text-transform: uppercase;
   }
+
 
 
   .heading{
@@ -108,6 +111,9 @@ background-size: cover;
 .inputWrap{
   position: relative;
 
+
+
+
   input{
     background-color: $dayVLightGray;
     border: none;
@@ -123,9 +129,7 @@ background-size: cover;
 
 
 
-.innerCircle{
-  @include circle(#fff, true)
-}
+
 }
 }
 
@@ -261,7 +265,7 @@ background-size: cover;
 
 .deskToggle{
 background-color:$dayVLightGray;
-transition: 2s;
+
 color:$dayVGrayBlue;
 display: flex;
 justify-content: space-around;
@@ -311,10 +315,8 @@ display:none;
       left: 0%;
       right:1rem;
     }
-    .innerCircle{
-      left: 28.75rem;
-     border-color:$dayVGrayBlue;
-    }
+
+
 }
 
 }
@@ -367,21 +369,16 @@ gap: 1rem;
 
 .darkModeBody{
   background-color:$nightVBlue;
-  transition: 2s;
+
 }
 .header{
 
   .inputWrap{
     .darkmodeInput{
       background-color: hsl(235, 24%, 19%);
-      transition: 2s;
+
       color:$dayVLightGray;
     }
-   .heading{
-    .innerCircle{
-      color:$dayVLightGray;
-    }
-   }
   }
 
 }
@@ -390,7 +387,7 @@ gap: 1rem;
   .darkmodeContent{
     background-color:$nightVDesBlue;
     color:$dayVLightGray;
-    transition: 2s;
+
   }
 
 }
